Rename leased SQS modules to coordinated
