+++ misk-aws2-sqs/src/main/kotlin/misk/aws2/sqs/jobqueue/coordinated/ForSqsReceiving.kt
@@ -1,4 +1,4 @@
-package misk.aws2.sqs.jobqueue.leased
+package misk.aws2.sqs.jobqueue.coordinated
 
 import jakarta.inject.Qualifier
