+++ misk-aws2-sqs/src/test/kotlin/misk/aws2/sqs/jobqueue/coordinated/QueueNameExtensionsTest.kt
@@ -1,4 +1,4 @@
-package misk.aws2.sqs.jobqueue.leased
+package misk.aws2.sqs.jobqueue.coordinated
 
 import kotlin.test.assertEquals
 import misk.aws2.sqs.jobqueue.DefaultDeadLetterQueueProvider
